pyproject.toml
[project]
name = "vector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "glfw>=2.9.0",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "pyopengl>=3.1.9",
    "pyopengl-accelerate>=3.1.9",
]
